+++ pocketsphinx/src/libpocketsphinx/hmm.c
@@ -55,7 +55,7 @@
 hmm_context_t *
 hmm_context_init(int32 n_emit_state,
 		 uint8 ** const *tp,
-		 int16 const *senscore,
+		 int32 const *senscore,
 		 uint16 * const *sseq)
 {
     hmm_context_t *ctx;
@@ -222,7 +222,7 @@ hmm_normalize(hmm_t *h, int32 bestscr)
 static int32
 hmm_vit_eval_5st_lr(hmm_t * hmm)
 {
-    int16 const *senscore = hmm->ctx->senscore;
+    int32 const *senscore = hmm->ctx->senscore;
     uint8 const *tp = hmm->ctx->tp[hmm->tmatid][0];
     uint16 const *sseq = hmm->senid;
     int32 s5, s4, s3, s2, s1, s0, t2, t1, t0, bestScore;
@@ -356,7 +356,7 @@ static int32
 hmm_vit_eval_5st_lr_mpx(hmm_t * hmm)
 {
     uint8 const *tp = hmm->ctx->tp[hmm->tmatid][0];
-    int16 const *senscore = hmm->ctx->senscore;
+    int32 const *senscore = hmm->ctx->senscore;
     uint16 * const *sseq = hmm->ctx->sseq;
     uint16 *ssid = hmm->senid;
     int32 bestScore;
@@ -529,7 +529,7 @@ hmm_vit_eval_5st_lr_mpx(hmm_t * hmm)
 static int32
 hmm_vit_eval_3st_lr(hmm_t * hmm)
 {
-    int16 const *senscore = hmm->ctx->senscore;
+    int32 const *senscore = hmm->ctx->senscore;
     uint8 const *tp = hmm->ctx->tp[hmm->tmatid][0];
     uint16 const *sseq = hmm->senid;
     int32 s3, s2, s1, s0, t2, t1, t0, bestScore;
@@ -610,7 +610,7 @@ static int32
 hmm_vit_eval_3st_lr_mpx(hmm_t * hmm)
 {
     uint8 const *tp = hmm->ctx->tp[hmm->tmatid][0];
-    int16 const *senscore = hmm->ctx->senscore;
+    int32 const *senscore = hmm->ctx->senscore;
     uint16 * const *sseq = hmm->ctx->sseq;
     uint16 *ssid = hmm->senid;
     int32 bestScore;
